Convert Gene Aliases to Official Gene Symbols

Maps gene alias names to official gene symbols.

Usage

alias2SymbolUsingNCBI(
  alias,
  gene.info.file,
  required.columns = c("GeneID", "Symbol", "description")
)

Arguments

alias

character vector of gene aliases

gene.info.file

the name of a gene information file downloaded from the NCBI.

required.columns

character vector of columns from the gene information file that are required in the output.

Value

A data frame with the latest gene symbols

See also

Examples

library(labyrinth)
update_gene_symbol('IL-6')
#> [1] "IL6"